Renamed component-folders and added fa for icons
