+++ src/assets/styles/_variables.scss
@@ -3,6 +3,8 @@ $xs: 599px;
 $sm: 600px;
 $md: 768px;
 $lg: 1200px;
+
+/* colors */
 $primaryDark: #212529;
 $primaryBlue: cadetblue;
 $lightGray: #fafafa;
