+++ src/views/home/components/styles.scss
@@ -1,0 +1,10 @@
+.icon-button {
+  background: none;
+  border: none;
+  cursor: pointer;
+  outline: none;
+
+  i {
+    margin-left: 5px;
+  }
+}
