+++ src/views/home/reducer.js
