+++ src/views/home/styles.scss
@@ -1,6 +1,10 @@
 @import "src/assets/styles/_variables.scss";
 
 .home-container {
+  h1 {
+    margin: 50px;
+  }
+
   .content-wrapper {
     display: flex;
     flex-wrap: wrap;
@@ -19,6 +23,12 @@
           padding: 0 50px 0 15px;
           font-size: 13px;
           min-width: 300px;
+        }
+      }
+
+      .suggestions {
+        ul {
+          list-style: none;
         }
       }
     }
